Put log stuff in a dedicated module. First step to something more generic
